backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
pymongo==4.10.1
python-dotenv==1.0.1
Scrapy==2.11.2
